+++ src/main/kotlin/com/teamsparta/buysell/domain/member/dto/request/SignUpRequest.kt
@@ -22,7 +22,7 @@ data class SignUpRequest (
         regexp = "^[\\w가-힣]{2,8}$",
         message = "닉네임은 숫자, 영어 대소문자, 한글을 포함한 최소 2자 최대 8자 이어야 하며, 중복, 공백은 허용하지 않습니다.")
     @Schema(description = "회원가입에 사용할 닉네임", example = "test123")
-    val nickname: String?,
+    val nickname: String,
 
     @field:Pattern(regexp = "^(남자|여자)$", message = "성별은 '남자' 또는 '여자' 중에서 선택해주세요.")
     val gender: String?,
