+++ src/main/kotlin/com/teamsparta/buysell/domain/member/model/Social.kt
@@ -7,11 +7,13 @@ import com.teamsparta.buysell.infra.auditing.SoftDeleteEntity
 import io.swagger.v3.oas.annotations.media.Schema
 import jakarta.persistence.*
 import org.hibernate.annotations.SQLDelete
+import org.hibernate.annotations.SQLRestriction
 
 @Entity
 @Table(name = "member")
 @Schema(description = "소셜 회원 정보")
 @SQLDelete(sql = "UPDATE member SET is_deleted = true WHERE id = ?")
+@SQLRestriction("is_deleted = false")
 class Social(
     @OneToOne(cascade = [CascadeType.ALL])
     @JoinColumn(name = "account_id")
