+++ src/main/kotlin/com/teamsparta/buysell/domain/member/repository/MemberRepository.kt
@@ -7,6 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 interface MemberRepository: JpaRepository<Member, Int> {
     fun findByEmail(email: String): Member?
     fun findByEmailAndPlatform(email:String, platform: Platform): Member?
-    fun findByNickname(nickname: String): Member?
-    fun findByEmailAndIsVerified(email: String, isVerified: Boolean): Member?
+    fun existsByNickname(nickname: String): Boolean
 }
