+++ src/main/kotlin/com/teamsparta/buysell/infra/redis/CacheConfig.kt
@@ -1,15 +1,12 @@
-package com.teamsparta.buysell.infra.cache
+/*
+package com.teamsparta.buysell.infra.redis
 
-import org.springframework.cache.CacheManager
 import org.springframework.cache.annotation.EnableCaching
-import org.springframework.cache.concurrent.ConcurrentMapCache
-import org.springframework.cache.support.SimpleCacheManager
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.redis.cache.RedisCacheConfiguration
 import org.springframework.data.redis.cache.RedisCacheManager
 import org.springframework.data.redis.connection.RedisConnectionFactory
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer
 import org.springframework.data.redis.serializer.RedisSerializationContext
 import org.springframework.data.redis.serializer.StringRedisSerializer
 import java.time.Duration
@@ -32,4 +29,4 @@ class CacheConfig {
             .fromConnectionFactory(redisConnectionFactory)
             .cacheDefaults(defaults).build()
     }
-}
+}*/
